Move benchmarker to utils
